requirement.txt
line-bot-sdk==2.4.2
Flask==2.3.2
geopy==2.3.0
scikit-learn==1.3.0
sqlite3  
